VendorMaster/requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
bidict==0.21.2
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
codegen==1.0
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.3
django-admin-index==1.4.0
django-allauth==0.43.0
django-apptemplates==1.5
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-choices==1.7.1
django-colorfield==0.3.2
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.4.0
django-jet==1.0.8
django-modeladmin-reorder==0.3.1
django-object-actions==3.0.1
django-ordered-model==3.4.1
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-registration==3.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
enum34==1.1.10
et-xmlfile==1.0.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.1.0
incremental==17.5.0
jdcal==1.4.1
kiteconnect==3.8.2
kombu==5.0.2
MarkupPy==1.14
msgpack==1.0.0
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
patterns==0.3
phonenumbers==8.12.12
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-engineio==3.6.0
python-memcached==1.59
python-socketio==4.0.3
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
raven==6.10.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
reversion==0.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.2
tornado==6.1
Twisted==20.3.0
txaio==20.4.1
Unipath==1.1
urllib3==1.25.11
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.2.0
